+++ src/LOCS/global_variables.h
@@ -31,9 +31,9 @@ extern struct spdk_nvme_ns * ns;
 extern struct spdk_nvme_detach_ctx *g_detach_ctx; //detach context
 
 // global varibales for define LOCS variables 
-extern uint32_t chunk_capacity; // maximum data entries of a chunk
+extern uint64_t chunk_capacity; // maximum data entries of a chunk
 
-extern uint64_t page_size;  // bytes of a page/sector
+extern uint64_t page_size;  // bytes of a page/(sector)
 
 extern struct channels_io *channels;
